solution of exercise 21.1 - revision 3 updates after code review - Thanks a lot :)
