+++ src/main/resources/static/styles.css
@@ -1,0 +1,86 @@
+html, body {
+    margin: 0;
+    padding: 0;
+    height: 100%;
+    font-family: sans-serif;
+}
+form {
+    display: grid;
+    grid-template-columns: 120px 1fr;
+}
+
+label {
+    grid-column: 1 / 2;
+    margin: 2px;
+    padding: 5px;
+}
+
+input {
+    border: 1px solid blue;
+    margin: 2px;
+    padding: 5px;
+    grid-column: 2 / 3;
+    max-width: 300px;
+}
+
+button {
+    border: 1px solid blue;
+    margin-top: 10px;
+    padding: 5px;
+    width: 100px;;
+    grid-column: 2 / 3;
+}
+select {
+    border: 1px solid blue;
+    margin: 2px;
+    padding: 5px;
+    grid-column: 2 / 3;
+    max-width: 312px;
+}
+thead {
+    text-align: center;
+}
+td {
+    border-bottom: 1px solid #ddd;
+    /*#ddd*/
+}
+
+/*tbody:nth-child(odd) {background-color: lightgray}*/
+#price-column {
+    text-align: center;
+    min-width: 200px;
+}
+#container {
+    min-height: 100%;
+    position: relative;
+    background: lightyellow;
+}
+
+#header {
+    background: indigo;
+    color: lightskyblue;
+    padding: 3px;
+    height: 20px;
+    font-weight: lighter;
+    text-align: center;
+}
+#link {
+    margin-top: 30px;
+}
+#middle {
+    background: lightyellow;
+    padding: 10px;
+    padding-bottom: 20px; /* Wysokość stopki */
+    padding-top: 20px; /* Wysokość nagłówka */
+}
+
+#footer {
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    height: 20px;
+    background: darkslategray;
+    color: lightskyblue;
+    text-align: right;
+    font-weight: lighter;
+}
