+++ src/main/resources/templates/commonElements.html
@@ -3,12 +3,12 @@
 
 <div th:fragment="Header">
     <div id="header">
-        <h6 class="text-center fw-lighter">solution for exercise 21.1</h6>
+        <span>solution for exercise 21.1</span>
     </div>
 </div>
 <div th:fragment="Footer">
     <div id="footer" class="text-right fw-lighter font-small blue">
-        last update: May 15th '22 by KUBAMAK
+        last update: May 18th '22 by KUBAMAK
     </div>
 </div>
 </html>
