--- src/main/resources/templates/formularz.html
@@ -1,77 +1,0 @@
-<!DOCTYPE html>
-<html lang="pl" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>BO:Dodaeanie artykułów</title>
-</head>
-<link
-        href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
-        rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
-        crossorigin="anonymous">
-<style>
-    html, body {
-        margin: 0;
-        padding: 0;
-        height: 100%;
-    }
-
-    #container {
-        min-height: 100%;
-        position: relative;
-        background: lightyellow;
-    }
-
-    #header {
-        background: indigo;
-        color: lightskyblue;
-        padding: 3px;
-        height: 20px;
-    }
-
-    #middle {
-        background: lightyellow;
-        padding: 10px;
-        padding-bottom: 20px; /* Wysokość stopki */
-        padding-top: 20px; /* Wysokość nagłówka */
-    }
-
-    #footer {
-        position: absolute;
-        bottom: 0;
-        width: 100%;
-        height: 20px;
-        background: darkslategray;
-        color: lightskyblue;
-    }
-</style>
-<body>
-<div id="container">
-    <div th:replace="commonElements::Header"></div>
-    <div id="middle">
-        <form th:action="@{/zapisz}" method="post" th:object="${product}" class="row g-3">
-            <h2 class="text-center"><br/>Bazar Obywateski - dodawanie artykułu<br/></h2>
-            <div class="col-md-6">
-                <label for="name" class="form-label fw-lighter">nazwa</label>
-                <input type="text" th:field="*{name}" class="form-control text-center" placeholder="nazwa">
-            </div>
-            <div class="col-md-2">
-                <label for="price" class="form-label fw-lighter">cena w PLN</label>
-                <input type="number" step="0.01" th:field="*{price}" class="form-control text-center" placeholder="1.25"><br/>
-            </div>
-            <div class="col-md-4">
-                <label for="category" class="form-label fw-lighter">kategoria</label>
-            <select name="category" id="category" class="form-control">
-                <option th:each="cat : ${T(pl.km.exercise211.Category).values()}"
-                        th:value="${cat}" th:text="${cat.description}"></option>
-            </select>
-            </div>
-            <div class="text-center">
-                <button type="submit" id="submit" class="btn btn-secondary">dodaj artykuł</button>
-            </div>
-        </form>
-        <a th:href="@{/}">Strona główna</a>
-    </div>
-    <div th:replace="commonElements::Footer"></div>
-</div>
-</body>
-</html>
