+++ src/main/resources/templates/index.html
@@ -4,46 +4,9 @@
     <meta charset="UTF-8">
     <title>Bazar Obywatelski</title>
 </head>
-<link
-        href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
-        rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
-        crossorigin="anonymous">
-<style>
-    html, body {
-        margin: 0;
-        padding: 0;
-        height: 100%;
-    }
 
-    #container {
-        min-height: 100%;
-        position: relative;
-        background: lightyellow;
-    }
+    <link rel="stylesheet" href="styles.css">
 
-    #header {
-        background: indigo;
-        color: lightskyblue;
-        padding: 3px;
-        height: 20px;
-    }
-
-    #middle {
-        background: lightyellow;
-        padding: 10px;
-        padding-bottom: 20px; /* Wysokość stopki */
-        padding-top: 20px; /* Wysokość nagłówka */
-    }
-
-    #footer {
-        position: absolute;
-        bottom: 0;
-        width: 100%;
-        height: 20px;
-        background: darkslategray;
-        color: lightskyblue;
-    }
-</style>
 <body>
 <div id="container">
     <div th:replace="commonElements::Header"></div>
@@ -52,11 +15,11 @@
         <h3>Bazar Obywatelski umożliwia przeglądanie przedmiotów wystawianych na sprzedaż przez innych użytkowników</h3><br/>
         <h4>Wybierz listę artykułów które cię interesują:</h4>
         <a th:href="@{|/produkty|}">Wszyskie</a><br/>
-        <a th:href="@{|/produkty?kategoria=${spozywcze.tag}|}"
+        <a th:href="@{/produkty(kategoria=${T(pl.km.exercise211.Category).GLOCERIES.tag})}"
            th:text="${T(pl.km.exercise211.Category).GLOCERIES.description}">Art. spożywcze</a><br/>
-        <a th:href="@{|/produkty?kategoria=${domowe.tag}|}"
+        <a th:href="@{/produkty(kategoria=${T(pl.km.exercise211.Category).HOUSEHOLD.tag})}"
            th:text="${T(pl.km.exercise211.Category).HOUSEHOLD.description}">Art. gosp. domowego</a><br/>
-        <a th:href="@{|/produkty?kategoria=${inne.tag}|}"
+        <a th:href="@{/produkty(kategoria=${T(pl.km.exercise211.Category).OTHER.tag})}"
            th:text="${T(pl.km.exercise211.Category).OTHER.description}">Inne</a><br/>
         <div>
             <br/>
