--- src/main/resources/templates/produkty.html
@@ -1,86 +1,0 @@
-<!DOCTYPE html>
-<html lang="pl" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Lista produktów</title>
-</head>
-<link
-        href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
-        rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3"
-        crossorigin="anonymous">
-<style>
-    html, body {
-        margin: 0;
-        padding: 0;
-        height: 100%;
-    }
-
-    #container {
-        min-height: 100%;
-        position: relative;
-        background: lightyellow;
-    }
-
-    #header {
-        background: indigo;
-        color: lightskyblue;
-        padding: 3px;
-        height: 20px;
-    }
-
-    #middle {
-        background: lightyellow;
-        padding: 10px;
-        padding-bottom: 20px; /* Wysokość stopki */
-        padding-top: 20px; /* Wysokość nagłówka */
-    }
-
-    #footer {
-        position: absolute;
-        bottom: 0;
-        width: 100%;
-        height: 20px;
-        background: darkslategray;
-        color: lightskyblue;
-    }
-</style>
-<body>
-<div id="container">
-    <div th:replace="commonElements::Header"></div>
-    <div id="middle">
-        <div th:unless="${category}">
-            <h1>Lista wszystkich artykułów</h1>
-        </div>
-        <div th:if="${category}">
-            <div th:switch="${category.tag}">
-                <h1 th:case="${T(pl.km.exercise211.Category).GLOCERIES.tag}"
-                    th:text="'Lista artykułów spożywczych'"></h1>
-                <h1 th:case="${T(pl.km.exercise211.Category).HOUSEHOLD.tag}"
-                    th:text="'Lista artykułów gospodarstwa domowego'"></h1>
-                <h1 th:case="${T(pl.km.exercise211.Category).OTHER.tag}"
-                    th:text="'Lista artykułów z kategorii inne'"></h1>
-            </div>
-        </div>
-        <table class="table">
-            <thead>
-            <tr>
-                <td>Nazwa</td>
-                <td>Cena w PLN</td>
-                <td>Kategoria</td>
-            </tr>
-            </thead>
-            <tbody>
-            <tr th:each="prod: ${products}">
-                <td th:text="${prod.name}"></td>
-                <td th:text="${prod.price}"></td>
-                <td th:text="${prod.category.description}"></td>
-            </tr>
-            </tbody>
-        </table>
-        <p th:text="|Suma cen wyświetlonych produktów: ${priceSum}zł|"></p>
-        <a th:href="@{/}">Strona główna</a>
-    </div>
-    <div th:replace="commonElements::Footer"></div>
-</div>
-</body>
-</html>
